Fine-tuned all of the Django models and did some testing so that everything seems to be in order.  More testing is needed, but at least it doesn't crash anymore.
Also, provided some sample code and a script that loads some data into
the database which you can access using the admin pages.  Customize the
'expect' script (the sample provided is server/nukeDB_sample.expect)
and it will handle the creation of the database and loading of some
toy data to play with.