Started adding "adapter" functionality to Shipyard (THIS IS NON-FUNCTIONAL).
So far we've created classes for raw inputs and outputs to Transformations,
and implemented a check on the input/output naming/numbering.  We still
need to fix up the cabling.  We are going to consult with Art to see
whether raw data should be allowed within a pipeline to be passed
between its steps.

Also, tests have not been implemented for the new stuff (nor for the last
round of additions, i.e. custom wiring, etc.).