Rewrote librarian tests for new ER setup.
Removed the test about an ER linking to a top-level pipeline, since that
can't happen anymore.

Changed the test about an ER linking to a sub-pipeline - instead of
linking to the pipeline pD, the ER now links directly to the only step
of that pipeline (namely step_D1).

Added a test for the type of general_transf(), because this was not
clear in the class diagram.