Minor code refactoring.
Moved messages.py to constants.py, and changed the corresponding imports.  Added the primary keys of the hard-coded Datatypes and CompoundDatatypes to constants.py.  Moved old test files in the sandbox app from tests_deprecated.py to deprecated_tests.py, etc., so that the new version of Django doesn't try to run them when you test the app.