Implemented prototype verification.
Yet another refactor of summarize_CSV. There is now a module-level
function called summarize_CSV in metadata.models, which takes as
input either a list of Datatypes, or a list of CDTM's.

Changed several helper functions in Datatype to only do their checks
once, the first time (and after that to recall the value from before).

Created helper _clean* functions for Datatype.clean(), since it was
very long.

Added module level loggers for Librarian and Metadata.

Added a parse_boolean function to Datatype.