Changes to Run* clean, is_complete, successful_execution finished.
Testing still remains.  There are definitely tests that fail right
now that will have to be addressed when the execute stuff is fixed.
In particular run_cable is definitely broken right now because
it doesn't handle the modifications to ExecLog, i.e. that ExecLogs
now know which record invoked them, not just which record they
belong to.

This addresses issues #44 and #49; we'll close those when the
testing is finished.  (Could be a while!)