Stopwatch class has been tested (via testing on a Run object).
Note that stopwatch/tests.py doesn't seem to work at all; either
this is a Django thing or I just haven't done it correctly.  I
suspect the former, as transformation/tests.py doesn't run either,
and there is a trivial test in that file.

This commit addresses and closes issue #40.