I wasn't sure where to put this, so I just made it up. Feel free to move this file to a more logical path.
This page demos some JS to filter a table based on its columns. Logical AND/OR gates and wildcards are implemented. No asynchronous communication with the server is needed, but may be implemented later if full-table sizes are unwieldly.

The intended use of this script is for our pipeline / data-set browser.