Fixed all broken tests resulting from the merge.
Most notable change is that the functionality that copied over
inputs/outputs from a parent revision when creating a new Method
has been moved from save() (where it had been commented out in
the master branch) to a new helper called copy_io_from_parent().
This is a better solution than putting it directly in the save()
function.