Fixed Pipeline[Family] deletion (closes #69).
The problem was the GenericRelation "execrecords" of Transformation.
This was a relic of the past, since not all Transformations can have
ExecRecords (only Methods). Removing this solved the problem. I tried
placing it in Method instead, but that only made it impossible to delete
a Method. So for now, it's commented out in Method.

Renamed create_complete() in ExecRecord to just create().

Added tests for deletion of various classes.

Made check in Pipeline's clean for consecutive indices cleaner.

Removed unnecessary details from documentation.

Removed some logging from execute_step.