Added tests for Run*.clean overchecking of outputs.
Decided not to move this functionality to ExecLog after all, since it is
more efficient to leave in RunAtomic. We have to go through all the
content and integrity checks there anyway, so we might as well just
check if one is duplicated at the same time.

Calling all_checks_passed() on all the RunAtomic's invoked logs also
turned out to be unnecessary. The call is made in RunAtomic.is_complete,
which assumes the RunAtomic is clean already. A clean RunAtomic can't
have any previous (ie. other invoked) ExecLogs which have not passed all
their checks. By the time we get to is_complete(), we already know the
invoked logs have passed all their checks and we don't have to check
again.

Closes #79.