Wrote is_identical comparator for Methods.
The function checks if two Methods have the same driver and the
same Transformation (ie. the same inputs and outputs). Wrote a few
unit tests for it.

Pipeline.is_identical() currently throws a NotImplementedError,
to prevent accidentally relying on Transformation.is_identical() for
Pipelines.

Made dataset_idx optional in Transformation.create_output, it now
defaults to the number of current outputs plus one.