Display correct file name with upload error.
Previously, when a Dataset was uploaded with an incorrect header,
an error would appear on the page claiming that the file "None"
had an incorrect header. This was due to previous modifications to
librarian functions to take file names or handles, instead of just
names. Fixed by checking whether we are working with a handle or a
file path, and displaying the name accordingly. Closes #178.