Minor change to execution algorithm and refactored execution code.
When executing steps and cables, the algorithm now checks after
the code has run whether any other ExecRecords for the same code
have been completed during the running.

All the tests run, but there is an issue with leaking file
descriptors.  The problem appears to be in the clean() routine
of CodeResourceRevision.