fixed an error in the MiCall demo coverage_plot (matplotlib initialisation) problem only presented itself when using ssh to run jobs.
modded the code in method.workers that invokes the method of a step to be more robust when
calling scripts without a shebang. Currently, method drivers without a shebang, are assumed
to be bash scripts. This will result in python drivers without a shebang to fail (kive reports
this, and bash's stderr can be examined).

Replaced an assert statement I was running into in fleet.workers.py with a logging output
and an exception for easier debugging.