added a check for a missing shebang in a method driver.
With Don, decided not to change the launching of method drivers, as
after experimenting, found that it is practically impossible to catch all
error cases. For example, when bash attempts to run a python script with
missing shebang, bash can simply hang forever.

used cherry-pick to fix error of missing FixtureFiles