First version of Kive using slurm.
a) added simplified instructions for slurm installation on ubuntu.

b) low level interface to slurm commands (sbatch, squeue and scancel) are
   implemented in the fleet.slurminterface.SlurmScheduler class.
   This class also reads slurm accounting information (in text form) about previously run jobs.

c) Implemented subclasses of WorkerInterface and ManagerInterface in slurminterface.py
   so that Kive can use slurm.
   Small changes were made to fleet.workers.Manager to accomodate slurm semantics: The
   ManagerInterface now has  worker_is_available() and a_task_is_finished() which replace
   probe_for_finished_worker().

e) Made changes in the method class so that a slurm job is submitted rather a call to ssh.
   Currently, slurm will run the job under the kive sandbox user account.

d) Added a new management command 'runslurm-manager' that starts a Manager with a
   SlurmInterface. This replaces the 'runfleet' command when using slurm.

Currently, all jobs are submitted with the same priority. Note that job priority
is enforced by the specific scheduler chosen by slurm configuration.
This has not been tested.

This commit also includes file in librarian/migrations which were missing from a previous
commit. These are files required due to the added date-time field for checking
the existence of external data sets.