The file Lookup function has been enhanced (provide more information) and also closes #624.
In addition, changes were made to the tabular view of run outputs because now log files
can be purged. A missing stdout file is no longer an indication of a failed run.
Instead, the return_code is used to determine a failed run.

In addition, the upload datasets from archive functionality has been enhanced and made more robust.