Volume is now being correctly removed.
Still need to figure out some things about the flow of the program,
including whether we need to explicitly check that the parent and
child processes are in the same process group.

Also need to figure out why test output from the clean up method
isn't appearing -- probably just because of the way output is being
redirected.