Restore Slurm node resource configuration
Slurm doesn't appear to check what resources nodes have, so if it
isn't configured, no jobs will be scheduled (as no nodes are
considered to have any CPU or memory). This commit restores the
configuration in the Vagrant hosts; the given configuration is
consistent with the VM specs in the Vagrantfile. We may wish to
reduce the amount of memory on each host if running both VMs on a
workstation becomes an issue.