Share data folder between head node and worker node
This commit alters Kive's Vagrant test environment to share a data
directory between the two nodes. This is necessary because jobs run
the Kive worker need access to input files placed in the
`media_root` directory by the Kive web app.

Because we're using a VirtualBox shared folder (which doesn't afford
altering directory permissions from the guest systems) the shared
directory has the most permissive permissions so that all uesrs on
both hosts can access them.

In a production environment, this function would be performed by a
network drive.