Update the PostGreSQL repository url
This commit updates the URL of the RPM that installs the PostGreSQL YUM
repository. Note that this isn't the URL for the RPM for PostGreSQL itself, but
for a metadata package which lets us do `yum install postgresql10-server`.

This update is necessary because the URL which it replaces stopped working.

The new URL is the one recommended by
https://www.postgresql.org/download/linux/redhat/