Allow database access from worker node
This commit contains changes to let processes on the worker nodes connect to
the Kive database (which is necessary for running Kive tasks). The changes
include:

- Adding a password to the Kive database user and adding it to the
  environment for the `kive` and `vagrant` users on both hosts.
- Configuring PostGres to use allow such connections
- Configuring the firewall to allow such connections, and to allow
  connections to the Kive test app from outside hosts

In the process of making these changes, additional files in
`vagrant_ubuntu` were consolidated to the `vagrant` environment, or
obviated. The `vagrant_ubuntu` directory has been preserved in its
existing state.