Use different CentOS Vagrant box
This commit changes the Vagrant box this environment is based on from
`centos/8` to `geerlingguy/centos8`. The former seems to have a bad DNF
configuration that makes it unable to find updates (specifically, it
seems to be trying to update using the Fedora project's mirror list,
which has a different API).