Add kive_server role
This commit adds a role for the Kive server, which pulls in the Kive
application code, web server, and database. It also includes fixes for problems
in other roles that were discovered while testing the kive_server role.

Furthermore, it removes some deployment-specific information from the Ansible
roles and factors them into a variables file, in preparation for supporting
multiple deployment configurations (e.g. a Dev configuration, a Test
configuration, and a Production configuration).