Add code for a development environment
This commit:

- Factors the Ansible roles created in the `cluster-setup` research project
  into their own top-level directory so that they can be shared between environments
- Adds a `dev-env` environment that uses the Ansible roles.

The environment includes a Vagrantfile for provisioning a suitable test host
(with Kive's source code mapped). It runs all of the components (head and
worker nodes for Kive and Slurm) in a single VM to reduce the memory and CPU
requirements for running it.

In addition to setting up Kive, the development environment includes a number of ergonimic features that aren't present in production:

- Automaticaly activate the Kive virtualenv and source environment variables
- Add the default `vagrant` user to the necessary groups
- Allow traffic on port 8080 (the default Kive development port)