Use Django's manifest static backend
This commit enables Django's
[`ManifestStaticFilesStorage`](https://docs.djangoproject.com/en/2.2/ref/contrib/staticfiles/#manifeststaticfilesstorage)
backend. This adds a step to the `collectstatic` management command which adds
the MD5 hash of a static file's contents to its path. This lets Kive set long
cache expiry times while avoiding the possibility of users having stale
JavaScript or CSS.

Files served using the
[`static`](https://docs.djangoproject.com/en/2.2/howto/static-files/) template
tag are treated in this fashion. Additionally, imports and URLs in CSS files
are parsed and replaced as appropriate.

User-uploaded files stored in the app's `MEDIA_ROOT` (in Kive's case,
singularity images and data files) aren't affected.