Catch empty zipfiles
This commit adds logic to catch empty Zipfiles when they're submitted as
containers. This case requires some special handling as `FieldField.file` that
gets passed to the container validation code doesn't exist, which causes
exceptions in Django's file validation code. The solution is to catch this case
before the brittler validation code runs and return early.