Test `purge.Command.get_file_size` behavior
This commit adds a failing test for `purge.Command`'s `get_file_size` method.
The desired behavhior (which isn't implemented) is for deleted files to be
treated as new files.

It also marks the `get_file_size` method as a `staticmethod` and removes a
deprecated `__future__` import that is redundant on Python 3.