Make purge treat missing files as changes
This commit alters the container app's `purge` task to treat missing files as
changed files (i.e. deleting a file from a directory will mark the directory as
active). This is to address a bug where purge and synch tasks were interfering
with each other (a file would be detected on disk but deleted before it could
be `stat`ed). The new behavior should allow the sync and purge tasks to
cooperate.