Fix broken tests
This commit fixes tests that were broken by the changes to how the
runcontainer module constructs its command. Some failing tests were
caused by under-specified arguments.

Furthermore, this commit changes the `argtype` property of
`ContainerArgument` to return `None` instead of raising a validation
error when a valid argument type can't be assigned. This is to avoid
raising an exception when accessing a field, which could be
surprising. The validation has been added to the model's `clean`
method instead.