Remove unneeded __future__ imports
This commit removes imports from the `__future__` module that are no
longer necessary after the upgrade to Python 3. Specifically, it
removes `print_function` and `unicode_literals`.

See the documentation of
[`__future__`](https://docs.python.org/3/library/__future__.html) for
more details.

It also removes annotations which force Emacs to treat source code as
UTF encoded. The [default encoding] of Python source code is now UTF8,
so editors can be configured to use UTF8 encoding for all `*.py`
files.

[default encoding]: https://docs.python.org/3/howto/unicode.html#the-string-type