Add a simple collation script
This commit adds a simple script for collating CSV files to the
Singularity example code. The purpose of this script is to serve as a
test-bed for the new multi-parameter input arguments; it probably
shouldn't be used to actually collate data in production.

An interesting etymological tidbit: the root of [collation] is a Latin
word meaning "the act of bringing together" which once also meant "a
light supper". This meaning is preserved in French, where
[collation][collation-fr] means "snack".

So anyone who came through French Immersion may be:

1. Transported back to their elementary school days.
2. Disappointed by the content of this app.

[collation]: https://www.etymonline.com/word/collation#etymonline_v_28313
[collation-fr]: https://dictionnaire.lerobert.com/definition/collation