Remove shared data directry from dev-env
This commit removes the shared Virtualbox directory from the dev-env
testing environment. For file-system and permissions reasons, the Kive
media root needs to be in the VM's filesystem (not a shared directory)
for this environment, and data can be persisted between VMs using the
default '/vagrant' directory. The shared directory is only a source of
confusion.