Special handling for input datasets of keyword-style arguments
This commit adds special handling of the files provided by input
datasets associated with keyword-style input arguments.

A `ContainerDataset` associates a dataset with a `ContainerArgument` for
one of the runs of that container app. While setting up the app run,
Kive copies the relevant data files into the run's sandbox.

The existing behavior is to always use the **argument's name** for the
sandbox file. With this change, that behavior is retained for fixed
position input files, but keyword-style arguments will use the filenames
that Kive assigns to the file.

This is necessary because the files given to a multi-input style keyword
argument need to be different (otherwise the input files just over-write
each other as they're copied into the sandbox).