Save output datasets with a unique filename
This commit adds a method to the `Dataset` class to create a unique name
based on the dataset's name and ID. It does this by adding the ID
between the file's name and extension (or at the end of the filename if
it has no extension). It alters the `runcontainer` command to use this
unique name for output datasets.