Generate Slurm config files from templates
This commit updates the `slurm_node` Ansible role to generate the
`slurm.conf` file from a template. This lets different environments
specify different numbers, and names of hosts while using the same
roles.

Currently, only the name and RAM of Slurm hosts is configurable, but
adding other fields to the template will be straightforward if it
becomes necessary.