Use absolute paths for output dir datasets
This commit adjusts the logic for naming the datasets that output
directory arguments to container runs create. To avoid depending on
invocation details (specifically, whether it's invoked with absolute or
relative paths), it converts all paths to absolute before stripping the
common leading directories from output arguments.

With this implementation, the output directory

    ${SANDBOX_ROOT}/outputdir/...

will generate datasets prefixed with `outputdir_`.