Add partial implementation of argument comparison
This commit adds partial implemention of a function to compare the
datasets associated with a re-run analysis and its original.

Prior to the addition of optional and multi-valued arguments, there was
always a one-to-one correspondence between app arguments and datasets.
Now that arguments may be missing or correspond to an arbitrary number
of inputs or outputs, there's no guarantee that this is the case. To
preserve existing functionality for comparing datasets between re-runs
and their original runs, a function to perform the comparison in each
case is required.

This commit includes the implementation for fixed inputs and outputs
(which mimics the current behavior), as well as setting up a datatype
for the comparison and test case (with substantial mocking
infrastructure) to test the functions.

The implementations of optional and and directory inputs will come in
subsequent commits.