Revert change in import/config order
This commit reverts a change that might cause problems with creating
Kive's WSGI Application object. Because Django must perform setup before
creating the object, any imports that depend on the
`DJANGO_SETTINGS_MODULE` variable might behave differently with the
changes that this commit reverts. The source of `get_wsgi_application`
looks like it should defer any relevant loading until after the variable
is set, but the change is easy to revert and has very limited scope.