Parametrize the slurmctld head node configuration
This commit sets the head node and account storage database node in the
`slurmctld` configuration file (`slurm.conf`) from an environment
variable. Previously, it was hard-coded.

This configuration is necessary because we want to use different head
nodes in different environments (`head` in the `dev-env`, `bulbasaur` or
`octomore` on the live clusters).